% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nodes.R
\name{compare_nodes}
\alias{compare_nodes}
\title{Compare Nodes}
\usage{
compare_nodes(node_1, node_2)
}
\arguments{
\item{node_1}{A vector representing the first node whose elements will be compared against the accolites of `node_2`.}

\item{node_2}{A vector representing the second node, from which accolites will be derived for comparison.}
}
\value{
A logical value. The function returns `TRUE` if all elements of `node_1` are found in the accolites of `node_2`. It returns `FALSE` otherwise.
}
\description{
This function compares two nodes to determine if all elements of the first node (`node_1`) are present within the accolites of the second node (`node_2`). Accolites are defined as a set of elements retrieved from the `get_accolites` function, which operates on `node_1` within a specified window of data. The comparison is successful if all elements in `node_1` are found in the accolites of `node_2`.
}
