% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_utilities.R
\name{.check_fits}
\alias{.check_fits}
\title{Check Fits}
\usage{
.check_fits(df, min_dist_motifs, len)
}
\arguments{
\item{df}{A data frame containing motif information with columns `start` and `end` indicating the positions of motifs.}

\item{min_dist_motifs}{A numeric value specifying the minimum distance required between motifs.}

\item{len}{A numeric value representing the total length of the curve.}
}
\value{
A logical value indicating whether the motifs fit within the curve and do not overlap.
}
\description{
This function checks if motifs fit within a specified curve length and whether there are overlaps between motifs. 
It returns TRUE if all conditions are met and FALSE otherwise.
}
\details{
Check Fits for Motifs
}
