% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_utilities.R
\name{.resample}
\alias{.resample}
\title{Resample Vector}
\usage{
.resample(x, ...)
}
\arguments{
\item{x}{A vector to be resampled.}

\item{...}{Additional arguments passed to the sampling function.}
}
\value{
A resampled vector.
}
\description{
This function resamples a vector by randomly selecting indices. The number of samples can be specified.
}
\details{
Resample Vector
}
