% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_utilities.R
\name{.transform_to_matrix}
\alias{.transform_to_matrix}
\title{Transform to Matrix}
\usage{
.transform_to_matrix(input)
}
\arguments{
\item{input}{A numeric vector or matrix to be transformed.}
}
\value{
A matrix representation of the input.
}
\description{
This function transforms a numeric vector into a matrix. If the input is a single number, it creates a 1x1 matrix. 
If the input is a vector, it creates a 1xN matrix. If the input is already a matrix, it returns it unchanged.
}
\details{
Transform Input to Matrix


- If the input is neither a numeric vector nor a matrix, an error is raised.
}
