% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_utilities.R
\name{generate_background_curve}
\alias{generate_background_curve}
\title{Generate Background Curve}
\usage{
generate_background_curve(len, dist_knots, norder, weights, add_noise)
}
\arguments{
\item{len}{An integer indicating the length of the curve to generate.}

\item{dist_knots}{A numeric value indicating the distance between knots.}

\item{norder}{An integer specifying the order of the B-spline.}

\item{weights}{A numeric vector containing the coefficients for the B-spline.}

\item{add_noise}{A logical value indicating whether to add Gaussian noise to the curve.}
}
\value{
A list containing the generated basis, coefficients, and curve values with or without noise.
}
\description{
This function generates a background curve using B-splines with specified parameters, including knots, order, and weights. 
Optionally, additive noise can be added to the curve.
}
\details{
Generate Background Curve
}
