% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_utilities.R
\name{generate_curve_vector}
\alias{generate_curve_vector}
\title{Generate Curve Vector from FD Object}
\usage{
generate_curve_vector(fd_curve, step_by = 1, Lfdobj = 0)
}
\arguments{
\item{fd_curve}{A functional data object from which to generate the curve.}

\item{step_by}{A numeric value indicating the step size for generating the curve.}

\item{Lfdobj}{A numeric value indicating the order of the derivative to compute. Default is 0 (no derivative).}
}
\value{
A numeric vector representing the generated curve.
}
\description{
This function generates a curve vector from a functional data (fd) object, with a specified step size and optional derivative.
}
\details{
Generate Curve Vector
}
