% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_node.R
\name{recommend_node}
\alias{recommend_node}
\title{Recommend Node from a Numeric Vector}
\usage{
recommend_node(node)
}
\arguments{
\item{node}{A numeric vector representing scores or metrics. Must contain at least two elements.}
}
\value{
An integer index indicating the recommended node based on the analysis of the input vector.
}
\description{
This function analyzes a numeric vector and recommends an index based on the characteristics of the vector. 
If the vector is strictly increasing, it suggests stopping before the maximum growth rate. 
If the vector is not strictly increasing, it recommends the index of the minimum value.
}
\details{
Recommend Node Function


- If the input vector has fewer than two elements, the function defaults to returning an index of 1.
- The function first checks if the input is a numeric vector. If not, an error is thrown.
- The function calculates the differences between successive elements and identifies whether the vector is strictly increasing.
- In the case of a strictly increasing vector, it returns the index of the maximum growth rate.
- For non-increasing vectors, it returns the index of the minimum value.
}
