% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedHighDimensional.R
\docType{data}
\name{simulatedHighDimensional}
\alias{simulatedHighDimensional}
\title{High Dimensional simulated data for local clustering}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(simulatedHighDimensional)
}
\description{
The functional dataset consists of 43 matrices for curves and an equal number 
for their derivatives of two columns and a variable number of rows ranging 
from a minimum of 1836 to a maximum of 43868. The data presents 
a lot of missing values (large gaps).
}
\keyword{datasets}
