% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\name{to_motifDiscovery,list-method}
\alias{to_motifDiscovery,list-method}
\title{to_motifDiscovery}
\usage{
\S4method{to_motifDiscovery}{list}(curves)
}
\arguments{
\item{curves}{A list coming from the generateCurves function.}
}
\value{
A list containing all curves formatted to be suitable for input into the discoverMotifs function.
}
\description{
Transforms the result of generateCurves into a format suitable for discoverMotifs
}
\examples{
\donttest{
builder <- funMoDisco::motifSimulationBuilder(N = 20, len = 300, mot_details = NULL)
curves <- funMoDisco::generateCurves(builder)
formatted_curves <- to_motifDiscovery(curves)
}
}
