% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\name{to_motifDiscovery}
\alias{to_motifDiscovery}
\title{to_motifDiscovery}
\usage{
to_motifDiscovery(curves)
}
\arguments{
\item{curves}{A list coming from the generateCurves function.}
}
\value{
A list containing all curves formatted to be suitable for input into the discoverMotifs function.
}
\description{
Transforms the result of generateCurves into a format suitable for discoverMotifs
}
\examples{
\donttest{
N <- 20 # number of curves
len <- 300 # length of the curves
mot_details <- NULL # or list()

builder <- motifSimulationBuilder(N = N, len = len, mot_details = NULL)
curves <- funMoDisco::generateCurves(builder)
formatted_curves <- to_motifDiscovery(curves)
}
}
