% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funMoDisco-pkg.R
\docType{package}
\name{funMoDisco-package}
\alias{funMoDisco}
\alias{funMoDisco-package}
\title{funMoDisco: Motif Discovery in Functional Data}
\description{
It provides tools for discovering functional motifs and clustering misaligned curves by implementing two complementary algorithms: ProbKMA and FunBIalign. ProbKMA (Cremona and Chiaromonte, 2020) is a probabilistic K-means algorithm that leverages local alignment and fuzzy clustering to identify recurring patterns or motifs across and within curves, allowing portions of curves to belong to multiple clusters. It efficiently handles the alignment of curve portions and adapts motif lengths in a data-driven manner, making it highly flexible for complex functional data. FunBIalign (Di Iorio, Cremona, and Chiaromonte, 2023) uses hierarchical agglomerative clustering with a functional generalization of the Mean Squared Residue Score to discover motifs in both single curves and sets of curves. This approach is particularly adept at identifying local structures and relationships among curves, with no need for multiple initializations or re-running the algorithm. The package also includes functionality to simulate functional data with embedded motifs, allowing users to create benchmark datasets to validate and compare motif discovery methods.
}
